requirements.txt
# pdf
fpdf2

# template
Jinja2==3.1.0